requirements.txt
Django~=4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.0
python-dotenv==1.0.0
gunicorn==21.0
psycopg2_binary==2.9.9
pillow==10.1.0
PyJWT==2.8.0
requests~=2.31.0